refactor: implement coderabbit suggestions
