+++ app/callback/route.ts
@@ -1,6 +1,6 @@
 import { handleAuth } from "@workos-inc/authkit-nextjs";
 import { NextResponse } from "next/server";
-import { isWorkOSEnabled } from "@/lib/auth-config";
+import { isWorkOSEnabled } from "@/lib/auth/client";
 
 // Redirect the user to `/` after successful sign in
 // The redirect can be customized: `handleAuth({ returnPathname: '/foo' })`
